LayoutTests/http/tests/workers/self-readonly-expected.txt

CONSOLE MESSAGE: NaN
Test passes if "NaN" is returned by the worker.
